# api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_florinel-chis_gophercrm_internal_models.Customer:
    properties:
      address:
        type: string
      city:
        type: string
      company:
        type: string
      country:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      notes:
        type: string
      phone:
        type: string
      position:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tickets:
        items:
          $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Ticket'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.User'
      user_id:
        type: integer
    type: object
  github_com_florinel-chis_gophercrm_internal_models.Lead:
    properties:
      classification:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.LeadClassification'
      company:
        type: string
      created_at:
        type: string
      customer:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Customer'
      customer_id:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      notes:
        type: string
      owner:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.User'
      owner_id:
        type: integer
      phone:
        type: string
      position:
        type: string
      source:
        type: string
      status:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.LeadStatus'
      updated_at:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.LeadClassification:
    enum:
    - unclassified
    - test
    - spam
    - lead
    - hot_lead
    type: string
    x-enum-varnames:
    - LeadClassificationUnclassified
    - LeadClassificationTest
    - LeadClassificationSpam
    - LeadClassificationLead
    - LeadClassificationHotLead
  github_com_florinel-chis_gophercrm_internal_models.LeadStatus:
    enum:
    - new
    - contacted
    - qualified
    - unqualified
    - converted
    type: string
    x-enum-varnames:
    - LeadStatusNew
    - LeadStatusContacted
    - LeadStatusQualified
    - LeadStatusUnqualified
    - LeadStatusConverted
  github_com_florinel-chis_gophercrm_internal_models.Task:
    properties:
      assigned_to:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.User'
      assigned_to_id:
        type: integer
      created_at:
        type: string
      customer:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Customer'
      customer_id:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      lead:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Lead'
      lead_id:
        type: integer
      priority:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TaskPriority'
      status:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TaskStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.TaskPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - TaskPriorityLow
    - TaskPriorityMedium
    - TaskPriorityHigh
  github_com_florinel-chis_gophercrm_internal_models.TaskStatus:
    enum:
    - pending
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusInProgress
    - TaskStatusCompleted
    - TaskStatusCancelled
  github_com_florinel-chis_gophercrm_internal_models.Ticket:
    properties:
      assigned_to:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.User'
      assigned_to_id:
        type: integer
      created_at:
        type: string
      customer:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Customer'
      customer_id:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TicketPriority'
      resolution:
        type: string
      status:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TicketStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.TicketPriority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - TicketPriorityLow
    - TicketPriorityMedium
    - TicketPriorityHigh
    - TicketPriorityUrgent
...
  github_com_florinel-chis_gophercrm_internal_models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      locked_until:
        type: string
      role:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.UserRole'
      updated_at:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.UserRole:
    enum:
    - admin
    - sales
    - support
    - customer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleSales
    - RoleSupport
    - RoleCustomer
  github_com_florinel-chis_gophercrm_internal_utils.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
      meta:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
      success:
        type: boolean
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.CreateTaskRequest:
    properties:
      assigned_to_id:
        type: integer
      customer_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      lead_id:
        type: integer
      priority:
        allOf:
        - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TaskPriority'
        enum:
        - low
        - medium
        - high
      title:
        type: string
    required:
    - assigned_to_id
    - title
    type: object
...
info:
  contact: {}
  description: CRM API for managing users, leads, customers, tickets, tasks and API
    keys.
  title: GopherCRM API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: List tasks. Admins see all tasks and may use search and sorting;
        every other role is silently narrowed to the tasks assigned to them, and the
        search, sort_by and sort_order parameters are ignored for them.
      parameters:
      - description: Page number (1-based); overrides offset when greater than 0
        in: query
        name: page
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - default: 20
        description: Page size (max 100)
        in: query
        name: limit
        type: integer
      - description: Sort column; ignored unless one of the allowed values, and ignored
          entirely for non-admin callers
        enum:
        - created_at
        - updated_at
        - title
        - status
        - priority
        - due_date
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction; anything else falls back to asc
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Free-text search across task fields; admin only, and takes precedence
          over sort_by
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tasks retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  properties:
                    tasks:
                      items:
                        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Task'
                      type: array
                    total:
                      type: integer
                  type: object
                meta:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task (admin, support and sales roles only). Non-admin
        callers may only assign the task to themselves; admins may assign it to any
        active user. A task may reference a lead or a customer, but not both. The
        new task always starts in status "pending" regardless of the request body.
      parameters:
      - description: Task creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Task'
              type: object
        "400":
          description: Invalid request data, assignee is deactivated, or task links
            both a lead and a customer
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin, support or sales role required, or non-admin
            assigning to another user
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "404":
          description: Assignee, lead or customer not found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create a new task
      tags:
      - tasks
...
  /tasks/my:
    get:
      description: List the tasks assigned to the authenticated user. Available to
        every authenticated role, including admins, who also see only their own tasks
        here. Sorting and search are not supported on this endpoint.
      parameters:
      - description: Page number (1-based); overrides offset when greater than 0
        in: query
        name: page
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - default: 20
        description: Page size (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  properties:
                    tasks:
                      items:
                        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Task'
                      type: array
                    total:
                      type: integer
                  type: object
                meta:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List tasks assigned to the current user
      tags:
      - tasks
...
securityDefinitions:
  ApiKeyAuth:
    description: API key issued via POST /api-keys. The plaintext key is shown only
      once at creation.
    in: header
    name: ApiKey
    type: apiKey
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"